[INFER] x64 fixes, bug 656920.
UltraBlame original commit: d3e08103053ba7bc56e02c4dc577cf9be114135c

+++ js/src/methodjit/StubCalls.cpp
@@ -2748,7 +2748,7 @@ stubs::TypeBarrierHelper(VMFrame &f, uint32 which)
     JS_ASSERT(which == 0 || which == 1);
 
     
-    Value &result = f.regs.sp[-1 - which];
+    Value &result = f.regs.sp[-1 - (int)which];
     result = f.regs.sp[0];
